requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
marshmallow==3.20.2
email-validator==2.1.0
python-dateutil==2.8.2
